+++ dnssd/attributes_test.go
@@ -7,39 +7,28 @@ import (
 )
 
 var _ = Describe("type Attributes", func() {
-	var attrs *Attributes
-
-	BeforeEach(func() {
-		attrs = &Attributes{}
-	})
-
-	Describe("func NewAttributes()", func() {
-		It("allows for a 'fluent' interface", func() {
-			attrs := NewAttributes().
-				Set("<key>", []byte("<value>")).
-				SetFlag("<flag>")
-
-			v, ok := attrs.Get("<key>")
-			Expect(v).To(Equal([]byte("<value>")))
-			Expect(ok).To(BeTrue())
-
-			Expect(attrs.HasFlags("<flag>")).To(BeTrue())
-		})
-	})
-
 	Context("binary attributes", func() {
-		Describe("func Set()", func() {
+		Describe("func WithPair()", func() {
 			It("sets the attribute", func() {
-				attrs.Set("<key>", []byte("<value>"))
+				attrs := NewAttributes().
+					WithPair("<key>", []byte("<value>"))
 
 				v, ok := attrs.Get("<key>")
 				Expect(v).To(Equal([]byte("<value>")))
 				Expect(ok).To(BeTrue())
 			})
 
+			It("does not modify the original attributes", func() {
+				attrs := NewAttributes()
+				attrs.WithPair("<key>", []byte("<value>"))
+
+				Expect(attrs.IsEmpty()).To(BeTrue())
+			})
+
 			It("replaces flags with the same key", func() {
-				attrs.SetFlag("<key>")
-				attrs.Set("<key>", []byte("<value>"))
+				attrs := NewAttributes().
+					WithFlag("<key>").
+					WithPair("<key>", []byte("<value>"))
 
 				_, ok := attrs.Get("<key>")
 				Expect(ok).To(BeTrue())
@@ -47,8 +36,9 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("replaces other binary attributes with the same key", func() {
-				attrs.Set("<key>", []byte("<value-1>"))
-				attrs.Set("<key>", []byte("<value-2>"))
+				attrs := NewAttributes().
+					WithPair("<key>", []byte("<value-1>")).
+					WithPair("<key>", []byte("<value-2>"))
 
 				v, ok := attrs.Get("<key>")
 				Expect(v).To(Equal([]byte("<value-2>")))
@@ -56,7 +46,8 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("allows association of a nil slice", func() {
-				attrs.Set("<key>", nil)
+				attrs := NewAttributes().
+					WithPair("<key>", nil)
 
 				v, ok := attrs.Get("<key>")
 				Expect(v).To(BeEmpty())
@@ -64,7 +55,8 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("allows association of an empty slice", func() {
-				attrs.Set("<key>", []byte{})
+				attrs := NewAttributes().
+					WithPair("<key>", []byte{})
 
 				v, ok := attrs.Get("<key>")
 				Expect(v).To(BeEmpty())
@@ -75,7 +67,8 @@ var _ = Describe("type Attributes", func() {
 				"it panics if the key is invalid",
 				func(k, expect string) {
 					Expect(func() {
-						attrs.Set(k, nil)
+						NewAttributes().
+							WithPair(k, nil)
 					}).To(PanicWith(MatchError(expect)))
 				},
 				Entry("empty", "", "key must not be empty"),
@@ -87,7 +80,8 @@ var _ = Describe("type Attributes", func() {
 
 		Describe("func Get()", func() {
 			It("returns the associated value", func() {
-				attrs.Set("<key>", []byte("<value>"))
+				attrs := NewAttributes().
+					WithPair("<key>", []byte("<value>"))
 
 				v, ok := attrs.Get("<key>")
 				Expect(v).To(Equal([]byte("<value>")))
@@ -95,19 +89,23 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("returns false if there is no such key", func() {
+				attrs := NewAttributes()
+
 				_, ok := attrs.Get("<key>")
 				Expect(ok).To(BeFalse())
 			})
 
 			It("returns false the key is a flag", func() {
-				attrs.SetFlag("<key>")
+				attrs := NewAttributes().
+					WithFlag("<key>")
 
 				_, ok := attrs.Get("<key>")
 				Expect(ok).To(BeFalse())
 			})
 
 			It("is case insensitive", func() {
-				attrs.Set("<KEY>", []byte("<value>"))
+				attrs := NewAttributes().
+					WithPair("<KEY>", []byte("<value>"))
 
 				v, ok := attrs.Get("<key>")
 				Expect(v).To(Equal([]byte("<value>")))
@@ -117,9 +115,10 @@ var _ = Describe("type Attributes", func() {
 
 		Describe("func Pairs()", func() {
 			It("returns the binary attributes", func() {
-				attrs.Set("<key-1>", []byte("<value-1>"))
-				attrs.Set("<key-2>", nil)
-				attrs.SetFlag("<key-3>")
+				attrs := NewAttributes().
+					WithPair("<key-1>", []byte("<value-1>")).
+					WithPair("<key-2>", nil).
+					WithFlag("<key-3>")
 
 				Expect(attrs.Pairs()).To(Equal(
 					map[string][]byte{
@@ -132,15 +131,25 @@ var _ = Describe("type Attributes", func() {
 	})
 
 	Context("flags", func() {
-		Describe("func SetFlag()", func() {
+		Describe("func WithFlag()", func() {
 			It("sets the flag", func() {
-				attrs.SetFlag("<key>")
+				attrs := NewAttributes().
+					WithFlag("<key>")
+
 				Expect(attrs.HasFlags("<key>")).To(BeTrue())
 			})
 
+			It("does not modify the original attributes", func() {
+				attrs := NewAttributes()
+				attrs.WithFlag("<key>")
+
+				Expect(attrs.IsEmpty()).To(BeTrue())
+			})
+
 			It("replaces binary attributes with the same key", func() {
-				attrs.Set("<key>", []byte("<value>"))
-				attrs.SetFlag("<key>")
+				attrs := NewAttributes().
+					WithPair("<key>", []byte("<value>")).
+					WithFlag("<key>")
 
 				_, ok := attrs.Get("<key>")
 				Expect(ok).To(BeFalse())
@@ -148,8 +157,10 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("has no effect if the flag is already set", func() {
-				attrs.SetFlag("<key>")
-				attrs.SetFlag("<key>")
+				attrs := NewAttributes().
+					WithFlag("<key>").
+					WithFlag("<key>")
+
 				Expect(attrs.HasFlags("<key>")).To(BeTrue())
 			})
 
@@ -157,7 +168,8 @@ var _ = Describe("type Attributes", func() {
 				"it panics if the key is invalid",
 				func(k, expect string) {
 					Expect(func() {
-						attrs.SetFlag(k)
+						NewAttributes().
+							WithFlag(k)
 					}).To(PanicWith(MatchError(expect)))
 				},
 				Entry("empty", "", "key must not be empty"),
@@ -171,8 +183,9 @@ var _ = Describe("type Attributes", func() {
 			DescribeTable(
 				"it returns the expected result",
 				func(expect bool, keys ...string) {
-					attrs.SetFlag("<key-1>")
-					attrs.SetFlag("<key-2>")
+					attrs := NewAttributes().
+						WithFlag("<key-1>").
+						WithFlag("<key-2>")
 
 					Expect(attrs.HasFlags(keys...)).To(Equal(expect))
 				},
@@ -184,13 +197,15 @@ var _ = Describe("type Attributes", func() {
 			)
 
 			It("returns false if the key has a binary value associated with it", func() {
-				attrs.Set("<key>", []byte("<value>"))
+				attrs := NewAttributes().
+					WithPair("<key>", []byte("<value>"))
 
 				Expect(attrs.HasFlags("<key>")).To(BeFalse())
 			})
 
 			It("is case insensitive", func() {
-				attrs.SetFlag("<KEY>")
+				attrs := NewAttributes().
+					WithFlag("<KEY>")
 
 				Expect(attrs.HasFlags("<key>")).To(BeTrue())
 			})
@@ -198,9 +213,10 @@ var _ = Describe("type Attributes", func() {
 
 		Describe("func Flags()", func() {
 			It("returns the flags that are set", func() {
-				attrs.SetFlag("<key-1>")
-				attrs.SetFlag("<key-2>")
-				attrs.Set("<key-3>", []byte("<value>"))
+				attrs := NewAttributes().
+					WithFlag("<key-1>").
+					WithFlag("<key-2>").
+					WithPair("<key-3>", []byte("<value>"))
 
 				Expect(attrs.Flags()).To(Equal(
 					map[string]struct{}{
@@ -211,60 +227,93 @@ var _ = Describe("type Attributes", func() {
 			})
 		})
 
-		Describe("func Delete()", func() {
+		Describe("func Without()", func() {
 			It("clears flags", func() {
-				attrs.SetFlag("<key>")
-				attrs.Delete("<key>")
+				attrs := NewAttributes().
+					WithFlag("<key>").
+					Without("<key>")
+
 				Expect(attrs.HasFlags("<key>")).To(BeFalse())
+			})
+
+			It("does not modify the original attributes", func() {
+				attrs := NewAttributes().
+					WithFlag("<key>")
+				attrs.Without("<key>")
+
+				Expect(attrs.HasFlags("<key>")).To(BeTrue())
 			})
 		})
 	})
 
 	Describe("func IsEmpty()", func() {
 		It("returns true if there are no attributes", func() {
+			attrs := NewAttributes()
 			Expect(attrs.IsEmpty()).To(BeTrue())
 
-			attrs.Set("<key>", []byte("<value>"))
-			attrs.SetFlag("<flag>")
-
-			attrs.Delete("<key>")
-			attrs.Delete("<flag>")
+			attrs = attrs.
+				WithPair("<key>", []byte("<value>")).
+				WithFlag("<flag>").
+				Without("<key>").
+				Without("<flag>")
 
 			Expect(attrs.IsEmpty()).To(BeTrue())
 		})
 
 		It("returns false if there are key/value pairs", func() {
-			attrs.Set("<key>", []byte("<value>"))
+			attrs := NewAttributes().
+				WithPair("<key>", []byte("<value>"))
+
 			Expect(attrs.IsEmpty()).To(BeFalse())
 		})
 
 		It("returns false if there are flags", func() {
-			attrs.SetFlag("<key>")
+			attrs := NewAttributes().
+				WithFlag("<key>")
+
 			Expect(attrs.IsEmpty()).To(BeFalse())
 		})
 	})
 
 	Context("TXT records", func() {
-		Describe("func FromTXT()", func() {
+		Describe("func WithTXT()", func() {
 			It("parses flags", func() {
-				ok, err := attrs.FromTXT("<KEY>")
+				attrs, ok, err := NewAttributes().
+					WithTXT("<KEY-1>")
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ok).To(BeTrue())
-				Expect(attrs.HasFlags("<key>")).To(BeTrue())
+
+				attrs, ok, err = attrs.
+					WithTXT("<KEY-2>")
+				Expect(err).ShouldNot(HaveOccurred())
+				Expect(ok).To(BeTrue())
+
+				Expect(attrs.HasFlags("<key-1>")).To(BeTrue())
+				Expect(attrs.HasFlags("<key-2>")).To(BeTrue())
 			})
 
 			It("parses binary attributes", func() {
-				ok, err := attrs.FromTXT("<KEY>=<value>")
+				attrs, ok, err := NewAttributes().
+					WithTXT("<KEY-1>=<value-1>")
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ok).To(BeTrue())
 
-				v, ok := attrs.Get("<key>")
-				Expect(v).To(Equal([]byte("<value>")))
+				attrs, ok, err = attrs.
+					WithTXT("<KEY-2>=<value-2>")
+				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ok).To(BeTrue())
+
+				Expect(attrs.Pairs()).To(Equal(
+					map[string][]byte{
+						"<key-1>": []byte("<value-1>"),
+						"<key-2>": []byte("<value-2>"),
+					},
+				))
 			})
 
 			It("parses binary attributes with empty values", func() {
-				ok, err := attrs.FromTXT("<KEY>=")
+				attrs, ok, err := NewAttributes().
+					WithTXT("<KEY>=")
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ok).To(BeTrue())
 
@@ -274,7 +323,8 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("ignores values that start with an equals character", func() {
-				ok, err := attrs.FromTXT("=<value>")
+				attrs, ok, err := NewAttributes().
+					WithTXT("=<value>")
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ok).To(BeFalse())
 
@@ -283,7 +333,8 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("ignores empty strings", func() {
-				ok, err := attrs.FromTXT("")
+				attrs, ok, err := NewAttributes().
+					WithTXT("")
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ok).To(BeFalse())
 
@@ -294,19 +345,32 @@ var _ = Describe("type Attributes", func() {
 			DescribeTable(
 				"it returns an error if the key is invalid",
 				func(pair, expect string) {
-					_, err := attrs.FromTXT(pair)
+					_, _, err := NewAttributes().
+						WithTXT(pair)
 					Expect(err).To(MatchError(expect))
 				},
 				Entry("contains UTF-8", "<ключ>", "invalid key '<ключ>', key must contain only printable ASCII characters"),
 				Entry("contains non-printable ASCII", "<\x00>", "invalid key '<\x00>', key must contain only printable ASCII characters"),
 			)
+
+			It("does not modify the original attributes", func() {
+				attrs := NewAttributes()
+
+				_, ok, err := attrs.
+					WithTXT("<key>=<value>")
+				Expect(err).ShouldNot(HaveOccurred())
+				Expect(ok).To(BeTrue())
+
+				Expect(attrs.IsEmpty()).To(BeTrue())
+			})
 		})
 
 		Describe("func ToTXT()", func() {
 			It("returns TXT record values containing key/value pairs as per RFC 6763", func() {
-				attrs.SetFlag("<key-1>")
-				attrs.Set("<key-2>", []byte("<value>"))
-				attrs.Set("<key-3>", nil)
+				attrs := NewAttributes().
+					WithFlag("<key-1>").
+					WithPair("<key-2>", []byte("<value>")).
+					WithPair("<key-3>", nil)
 
 				Expect(attrs.ToTXT()).To(ConsistOf(
 					"<key-1>",
@@ -316,10 +380,11 @@ var _ = Describe("type Attributes", func() {
 			})
 
 			It("always places the 'version tag' attribute at the beginning, and sorts the other keys lexically", func() {
-				attrs.SetFlag("<key-3>")
-				attrs.Set("<key-2>", []byte("<value>"))
-				attrs.Set("<key-1>", nil)
-				attrs.Set("txtvers", []byte("1"))
+				attrs := NewAttributes().
+					WithFlag("<key-3>").
+					WithPair("<key-2>", []byte("<value>")).
+					WithPair("<key-1>", nil).
+					WithPair("txtvers", []byte("1"))
 
 				// Repeat test several times to ensure it's not just passing due
 				// to Go's psuedo-random map ordering.