+++ dnssd/unicastresolver_test.go
@@ -32,10 +32,12 @@ var _ = Context("UnicastResolver", func() {
 			TargetPort:  12345,
 			Priority:    10,
 			Weight:      20,
-			Attributes:  []Attributes{{}},
-			TTL:         DefaultTTL / 2, // Use something other than the default for LookupInstance() test
+			Attributes: []Attributes{
+				NewAttributes().
+					WithPair("<key>", []byte("<instance-a>")),
+			},
+			TTL: DefaultTTL / 2, // Use something other than the default for LookupInstance() test
 		}
-		instanceA.Attributes[0].Set("<key>", []byte("<instance-a>"))
 
 		instanceB = ServiceInstance{
 			Instance:    "Instance B",
@@ -45,10 +47,13 @@ var _ = Context("UnicastResolver", func() {
 			TargetPort:  12345,
 			Priority:    10,
 			Weight:      20,
-			Attributes:  []Attributes{{}, {}},
+			Attributes: []Attributes{
+				NewAttributes().
+					WithPair("<key>", []byte("<instance-b0>")),
+				NewAttributes().
+					WithPair("<key>", []byte("<instance-b1>")),
+			},
 		}
-		instanceB.Attributes[0].Set("<key>", []byte("<instance-b0>"))
-		instanceB.Attributes[1].Set("<key>", []byte("<instance-b1>"))
 
 		instanceC = ServiceInstance{
 			Instance:    "Instance C",