+++ dnssd/unicastserver_test.go
@@ -30,9 +30,11 @@ var _ = Context("UnicastServer", func() {
 			TargetPort:  12345,
 			Priority:    10,
 			Weight:      20,
-			Attributes:  []Attributes{{}},
+			Attributes: []Attributes{
+				NewAttributes().
+					WithPair("<key>", []byte("<instance-a>")),
+			},
 		}
-		instanceA.Attributes[0].Set("<key>", []byte("<instance-a>"))
 
 		instanceB = ServiceInstance{
 			Instance:    "Instance B",
@@ -42,10 +44,13 @@ var _ = Context("UnicastServer", func() {
 			TargetPort:  12345,
 			Priority:    10,
 			Weight:      20,
-			Attributes:  []Attributes{{}, {}},
+			Attributes: []Attributes{
+				NewAttributes().
+					WithPair("<key>", []byte("<instance-b0>")),
+				NewAttributes().
+					WithPair("<key>", []byte("<instance-b1>")),
+			},
 		}
-		instanceB.Attributes[0].Set("<key>", []byte("<instance-b0>"))
-		instanceB.Attributes[1].Set("<key>", []byte("<instance-b1>"))
 
 		instanceC = ServiceInstance{
 			Instance:    "Instance C",